services/orchestrator-service/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
pydantic-settings==2.5.2
python-multipart==0.0.9
pandas==2.2.3
openpyxl==3.1.5
pytest==8.3.3
httpx==0.27.2
python-docx==1.2.0
